Extract delete dialog
